refactor(web, exception): Web 패키지 구조 변경 - 구성 요소들을 web 패키지 아래에 수평적 구조로 구성 - exception 패키지 분리
