+++ src/main/java/app/handler/RegisterHandlerImpl.java
@@ -4,8 +4,8 @@ import app.db.Database;
 import app.model.User;
 import http.HttpMethod;
 import http.request.HttpRequest;
-import web.web.handler.DynamicViewHandler;
-import web.web.handler.response.view.ViewResponse;
+import web.handler.DynamicViewHandler;
+import web.response.ViewResponse;
 
 public class RegisterHandlerImpl extends DynamicViewHandler {
     public RegisterHandlerImpl() {
