+++ src/main/java/bootstrap/WebServer.java
@@ -1,4 +1,4 @@
-package web;
+package bootstrap;
 
 import java.net.ServerSocket;
 import java.net.Socket;
