+++ src/main/java/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package web.exception;
+package exception;
 
 
 import http.HttpStatus;
