+++ src/main/java/exception/ExceptionHandlerMapping.java
@@ -1,4 +1,4 @@
-package web.exception;
+package exception;
 
 import java.net.Socket;
 import java.util.List;
