+++ src/main/java/web/posthandler/ViewResponseHandler.java
@@ -1,11 +1,11 @@
-package web.web.handler.response.handler;
+package web.posthandler;
 
 import config.VariableConfig;
-import web.exception.ErrorException;
+import exception.ErrorException;
 import http.HttpStatus;
 import http.response.HttpResponse;
-import web.web.handler.response.WebHandlerResponse;
-import web.web.handler.response.view.ViewResponse;
+import web.response.WebHandlerResponse;
+import web.response.ViewResponse;
 
 import java.io.BufferedInputStream;
 import java.io.File;
