+++ src/test/java/web/handler/StaticContentHandlerTest.java
@@ -1,4 +1,4 @@
-package web.web.handler;
+package web.handler;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
